requirements.txt
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.3.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
bleach==3.1.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
configparser==4.0.2
contextlib2==0.6.0.post1
coverage==5.0.3
cryptography==2.1.4
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
enum34==1.1.6
fastavro==0.22.9
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.18.2
futures==3.3.0
idna==2.8
importlib-metadata==1.4.0
ipaddress==1.0.23
ipykernel==4.10.1
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
Jinja2==2.10.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==5.2.0
jupyter-core==4.6.1
jupyterlab==0.33.12
jupyterlab-launcher==0.11.2
keyring==10.6.0
keyrings.alt==3.0
MarkupSafe==1.1.1
mistune==0.8.4
mock==3.0.5
more-itertools==5.0.0
Naked==0.1.31
nbconvert==5.6.1
nbformat==5.0.3
notebook==5.7.8
packaging==20.0
pandocfilters==1.4.2
pathlib2==2.3.5
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.13.1
prettytable==0.7.2
prometheus-client==0.7.1
prompt-toolkit==1.0.18
ptyprocess==0.6.0
py==1.8.1
py4j==0.10.7
pycparser==2.19
pycrypto==2.6.1
Pygments==2.5.2
pygobject==3.26.1
pynput==1.6.5
pyparsing==2.4.6
pyrsistent==0.15.6
pyspark==2.4.4
python-dateutil==2.8.1
python-xlib==0.26
pyxdg==0.25
PyYAML==5.3
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
scandir==1.10.0
SecretStorage==2.3.1
Send2Trash==1.5.0
shellescape==3.4.1
simple-crypt==4.1.7
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.14.0
subprocess32==3.5.4
terminado==0.8.3
testpath==0.4.4
tornado==5.1.1
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==1.0.0
